Move MapInfoWindowShownEvent to Event package
