+++ app/src/main/java/lt/vilnius/tvarkau/fragments/BaseMapFragment.java
@@ -2,7 +2,6 @@ package lt.vilnius.tvarkau.fragments;
 
 
 import android.os.Bundle;
-import android.util.Log;
 
 import com.google.android.gms.maps.CameraUpdateFactory;
 import com.google.android.gms.maps.GoogleMap;
@@ -19,7 +18,7 @@ import java.util.HashMap;
 
 import lt.vilnius.tvarkau.R;
 import lt.vilnius.tvarkau.entity.Problem;
-import lt.vilnius.tvarkau.factory.MapInfoWindowShownEvent;
+import lt.vilnius.tvarkau.events.MapInfoWindowShownEvent;
 import lt.vilnius.tvarkau.views.adapters.MapsInfoWindowAdapter;
 
 /**
